+++ app/server/packages/VideoPackage.js
@@ -130,7 +130,8 @@ Object.freeze(VideoPackage.stateMachine);
 /**
  * Generates a thumbnail for the video.
  *
- * It uses ffmpeg to extract an image from the video.
+ * If no thumbnail has been provided by the user form, ffmpeg will be
+ * used to extract an image from the video to generate a thumbnail.
  *
  * This is a transition.
  *
@@ -140,25 +141,48 @@ VideoPackage.prototype.generateThumb = function() {
   var self = this;
   var filePath = this.getMediaFilePath();
 
-  // Generate thumb
   this.updateState(this.mediaPackage.id, STATES.GENERATE_THUMB, function() {
     var destinationPath = path.join(self.publishConf.videoTmpDir, String(self.mediaPackage.id));
-    process.logger.debug('Generate thumbnail (' + self.mediaPackage.id + ') in ' + destinationPath);
-    ffmpeg(filePath).screenshots({
-      timestamps: ['10%'],
-      filename: 'thumbnail.jpg',
-      folder: destinationPath
-    }).on('error', function(error) {
-      self.setError(new VideoPackageError(error.message, ERRORS.GENERATE_THUMB));
-    }).on('end', function() {
-      self.videoModel.updateThumbnail(
-        self.mediaPackage.id,
-        '/publish/' + self.mediaPackage.id + '/thumbnail.jpg',
-        function() {
-          self.fsm.transition();
+
+    if (self.mediaPackage.originalThumbnailPath !== undefined) {
+    // Copy user's thumbnail
+      process.logger.debug('Copy thumbnail (' + self.mediaPackage.id + ') in ' + destinationPath);
+      openVeoApi.fileSystem.copy(
+        self.mediaPackage.originalThumbnailPath,
+        path.join(destinationPath, 'thumbnail.jpg'),
+        function(error) {
+          if (error) {
+            self.setError(new VideoPackageError(error.message, ERRORS.COPY_THUMB));
+          }
+
+          self.videoModel.updateThumbnail(
+            self.mediaPackage.id,
+            '/publish/' + self.mediaPackage.id + '/thumbnail.jpg',
+            function() {
+              self.fsm.transition();
+            }
+          );
         }
       );
-    });
+    } else {
+    // Generate thumb
+      process.logger.debug('Generate thumbnail (' + self.mediaPackage.id + ') in ' + destinationPath);
+      ffmpeg(filePath).screenshots({
+        timestamps: ['10%'],
+        filename: 'thumbnail.jpg',
+        folder: destinationPath
+      }).on('error', function(error) {
+        self.setError(new VideoPackageError(error.message, ERRORS.GENERATE_THUMB));
+      }).on('end', function() {
+        self.videoModel.updateThumbnail(
+          self.mediaPackage.id,
+          '/publish/' + self.mediaPackage.id + '/thumbnail.jpg',
+          function() {
+            self.fsm.transition();
+          }
+        );
+      });
+    }
   });
 };